s3_config should be in lib.config
