+++ script/upload-checksums.py
@@ -5,7 +5,8 @@ import hashlib
 import os
 import tempfile
 
-from lib.util import download, rm_rf, s3_config, s3put
+from lib.config import s3_config
+from lib.util import download, rm_rf, s3put
 
 
 DIST_URL = 'https://atom.io/download/atom-shell/'
